Move all the new code to a new view
Move all the new code to a new view called ViewStructVSClass
