+++ CWC-Receta-Challenges/CWC-Receta-Challenges.xcodeproj/xcuserdata/mgonzalez506.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Bucket
+   uuid = "F8DC3F6E-FFFB-4DE9-8E2C-EEC1A89EC23A"
+   type = "1"
+   version = "2.0">
+</Bucket>
